requirements.txt
sentence-transformers
langchain
streamlit
docx2txt